pyproject.toml
[project]
name = "adk-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.15.1",
    "praw>=7.8.1",
    "python-dotenv>=0.21.0",
]
